Add background skillProficiencies
